changed id to long

+++ src/main/java/com/jore/epoc/dto/CompletedUserSimulationDto.java
@@ -6,8 +6,8 @@ import lombok.Data;
 
 @Data
 public class CompletedUserSimulationDto implements DataTransferObject {
-    private Integer simulationId;
+    private Long simulationId;
     private String simulationName;
     private String companyName;
-    private Integer companyId;
+    private Long companyId;
 }
