+++ src/main/java/com/jore/epoc/repositories/CompanyRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.jore.epoc.bo.Company;
 
-public interface CompanyRepository extends CrudRepository<Company, Integer> {
+public interface CompanyRepository extends CrudRepository<Company, Long> {
 }
