+++ src/main/java/com/jore/epoc/repositories/CompanySimulationStepRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.jore.epoc.bo.step.CompanySimulationStep;
 
-public interface CompanySimulationStepRepository extends CrudRepository<CompanySimulationStep, Integer> {
+public interface CompanySimulationStepRepository extends CrudRepository<CompanySimulationStep, Long> {
 }
