+++ src/main/java/com/jore/epoc/repositories/SettingRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.jore.epoc.bo.settings.EpocSetting;
 
-public interface SettingRepository extends CrudRepository<EpocSetting, Integer> {
+public interface SettingRepository extends CrudRepository<EpocSetting, Long> {
     Optional<EpocSetting> findBySettingKey(String key);
 }
