+++ src/main/java/com/jore/epoc/repositories/SimulationRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.jore.epoc.bo.Simulation;
 
-public interface SimulationRepository extends CrudRepository<Simulation, Integer> {
-    List<Simulation> findByIsStartedAndOwnerId(boolean isStarted, Integer id);
+public interface SimulationRepository extends CrudRepository<Simulation, Long> {
+    List<Simulation> findByIsStartedAndOwnerId(boolean isStarted, Long id);
 }
