+++ src/main/java/com/jore/epoc/repositories/SimulationStepRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.jore.epoc.bo.step.SimulationStep;
 
-public interface SimulationStepRepository extends CrudRepository<SimulationStep, Integer> {
+public interface SimulationStepRepository extends CrudRepository<SimulationStep, Long> {
 }
