+++ src/main/java/com/jore/epoc/repositories/UserInCompanyRoleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.jore.epoc.bo.user.UserInCompanyRole;
 
-public interface UserInCompanyRoleRepository extends CrudRepository<UserInCompanyRole, Integer> {
+public interface UserInCompanyRoleRepository extends CrudRepository<UserInCompanyRole, Long> {
     Iterable<UserInCompanyRole> findByIsInvitationRequired(boolean isInvitationRequired);
 }
